payment 파일 컴포넌트화 시킴
